Area(s) for Improvement
One area for improvement is trying to condense your work, I noticed in the first few code chunks there were some code that did the same thing- reading in the data twice. This is a well done script and I think it just needs a once over to check that it is as efficient as possible.