Areas of Strength
I like how you separated each thing you were doing with their own labels, it makes looking back at the data a lot easier as well as the comments within code are helpful for understanding what the code means as well.
The use of Facet_wrap() in your graphs was a good choice because it condensed a lot of information into one graph.
I like the use of your descriptives table and including relevant means all in one table. It acts as an easy visual reference for a lot of variables that otherwise wouldn't be easy to compare side by side.